+++ icx-cert/src/pprint.rs
@@ -14,7 +14,7 @@ struct StructuredCertHeader<'a> {
 /// A fully parsed replica certificate.
 #[derive(Deserialize)]
 struct ReplicaCertificate {
-    tree: HashTree<'static>,
+    tree: HashTree,
     signature: serde_bytes::ByteBuf,
 }
 
@@ -118,7 +118,7 @@ pub fn pprint(url: String, accept_encodings: Option<Vec<String>>) -> Result<()> 
     );
     println!("TREE HASH: {}", hex::encode(tree.digest()));
     println!("SIGNATURE: {}", hex::encode(cert.signature.as_ref()));
-    if let LookupResult::Found(mut date_bytes) = cert.tree.lookup_path(&["time".into()]) {
+    if let LookupResult::Found(mut date_bytes) = cert.tree.lookup_path(&["time"]) {
         let timestamp_nanos = leb128::read::unsigned(&mut date_bytes)
             .with_context(|| "failed to decode certificate time as LEB128")?;
         let dt = OffsetDateTime::from_unix_timestamp_nanos(timestamp_nanos as i128)